// apps/web/src/enums.ts
export enum OptmisticPublicationType {
    NewPost = 'NEW_POST',
    NewComment = 'NEW_COMMENT'
  }
  
  export enum NotificationType {
    All = 'ALL',
    Mentions = 'MENTIONS',
    Comments = 'COMMENTS',
    Likes = 'LIKES',
    Collects = 'COLLECTS'
  }
  
  export enum ProfileFeedType {
    Feed = 'FEED',
    Replies = 'REPLIES',
    Media = 'MEDIA',
    Collects = 'COLLECTS',
    Nft = 'NFT'
  }